test(element): move HTMLElementMock to a separate module
